backend/requirements.txt
# --- Core Django ---
Django>=5.0
djangorestframework>=3.15
python-dotenv>=1.0.1

# --- PDF Processing ---
PyPDF2>=3.0.1

# --- LangChain + Google Gemini ---
langchain>=0.2.0
langchain-community>=0.2.0
langchain-google-genai>=0.0.12
google-generativeai>=0.7.2

# --- Vector Database ---
faiss-cpu>=1.8.0

# --- Utilities ---
uvicorn>=0.30.0   # if you want to run with ASGI (fast dev server)
